\name{chgDefaults-methods}
\docType{methods}
\alias{chgDefaults-methods}
\alias{chgDefaults,ANY-method}
\alias{chgDefaults,psi_func-method}
\title{Methods for Function chgDefaults() in Package `robustbase'}
\description{
  This is auxiliary functionality for modifying
  \code{\link{psi_func-class}} objects.
}
\section{Methods}{
  \describe{
    \item{\code{object = "psi_func"}}{The method is used to change the default
      values for the tuning parameters, and returns a
      \code{\link{psi_func-class}} object, a copy of input \code{object}
      with the slot \code{tDefs} possibly changed;.
    }
  }
}
\examples{
## FIXME -- add one %% --> ../experi-psi-rho-funs.R
}
\keyword{methods}

