\name{fitted.nlrob}
\alias{fitted.nlrob}

\title{ Extract Model Fitted Values }
\description{
  \code{fitted} is a generic function which extracts fitted values from
  objects returned by modeling functions. \code{fitted.values} is an
  alias for it. 
}

\usage{
fitted.nlrob(object, ...)
}

\arguments{
  \item{object}{ nlrob object }
  \item{\dots}{ additional arguments }
}


\value{
  Fitted values extracted from \code{object}.
}

\author{ Andreas Ruckstuhl }

\seealso{ \code{\link[stats]{fitted}}}

\keyword{ regression }
