\name{lmrob.fit.MM}
\alias{lmrob.fit.MM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ MM-regression estimators  }
\description{
  Fit function for MM-regression estimators
}
\usage{
lmrob.fit.MM()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\details{ This function is used by \code{lmrob} and not
intended to be used on its own}
\value{
  A list that contains regression coefficient estimators
and the robustly estimated error standard deviation
}
\seealso{ \code{\link{lmrob}} }
\keyword{ robust }% at least one, from doc/KEYWORDS
\keyword{ regression }% __ONLY ONE__ keyword per line

